refactor(server): complete templates import hygiene; sweep imports to nodes/*; move tests and update to lifecycle-only; remove legacy references (#211)
