+++ apps/server/src/agents/simple.agent.ts
@@ -13,8 +13,8 @@ import { CheckpointerService } from '../services/checkpointer.service';
 import { ConfigService } from '../services/config.service';
 import { LoggerService } from '../services/logger.service';
 import { BaseAgent } from './base.agent';
-import { BaseTool } from '../tools/base.tool';
-import { LangChainToolAdapter } from '../tools/langchainTool.adapter';
+import { BaseTool } from '../nodes/tools/base.tool';
+import { LangChainToolAdapter } from '../nodes/tools/langchainTool.adapter';
 import { SummarizationNode } from '../nodes/summarization.node';
 import { NodeOutput } from '../types';
 import { z } from 'zod';
