+++ apps/server/src/nodes/agent.node.ts
@@ -12,8 +12,8 @@ import { ToolsNode } from './tools.node';
 import { CheckpointerService } from '../services/checkpointer.service';
 import { ConfigService } from '../services/config.service';
 import { LoggerService } from '../services/logger.service';
-import { BaseTool } from '../tools/base.tool';
-import { LangChainToolAdapter } from '../tools/langchainTool.adapter';
+import { BaseTool } from './tools/base.tool';
+import { LangChainToolAdapter } from './tools/langchainTool.adapter';
 import { SummarizationNode } from './summarization.node';
 import { EnforceRestrictionNode } from './enforceRestriction.node';
 import { NodeOutput } from '../types';
@@ -21,7 +21,7 @@ import { z } from 'zod';
 import { stringify as toYaml } from 'yaml';
 import { buildMcpToolError } from '../mcp/errorUtils';
 import { withAgent } from '@hautech/obs-sdk';
-import { TriggerListener, TriggerMessage, isSystemTrigger } from '../triggers/base.trigger';
+import { TriggerListener, TriggerMessage, isSystemTrigger } from './triggers/base.trigger.node';
 import { JSONSchema } from 'zod/v4/core';
 import { MessagesBuffer, ProcessBuffer } from '../agents/messages-buffer';
